Feature #46: connect navigation

+++ feature/report/src/main/java/com/hankki/feature/report/model/LocationModel.kt
@@ -1,8 +1,8 @@
 package com.hankki.feature.report.model
 
 data class LocationModel(
-    val latitude: Double = 0.0,
-    val longitude: Double = 0.0,
+    val latitude: Float = 0f,
+    val longitude: Float = 0f,
     val location: String = "",
     val address: String = ""
 )
